Create details page componemt
