+++ src/Pages/DetailsPage.js
@@ -1,0 +1,7 @@
+import React from 'react';
+
+const DetailsPage = () => (
+  <div>DetailsPage</div>
+);
+
+export default DetailsPage;
